+++ docs/.vuepress/config.js
@@ -58,10 +58,4 @@ module.exports = {
       }
     }
   }
-  // postcss: {
-  //   plugins: [
-  //     require('css-prefers-color-scheme/postcss'),
-  //     require('autoprefixer')
-  //   ]
-  // }
 }